requirements.txt
camelot-py[cv]
matplotlib
pandas
numpy